backend/requirements.txt
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11 # usado apenas se trocar para PostgreSQL
PyJWT==2.10.1
python-decouple==3.8
